requirements.txt
# Core dependencies for acoustic simulation framework
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.11.0
pyyaml>=6.0
pyroomacoustics>=0.7.0

# Optional dependencies for advanced features
matplotlib>=3.4.0
pandas>=1.3.0
tqdm>=4.62.0
